Secure the application

+++ app/api/issues/[id]/route.ts
@@ -4,11 +4,17 @@ import prisma from '@/prisma/client';
 
 import { issueSchema } from '@/app/validationSchema';
 
+import { getServerSession } from 'next-auth';
+import authOptions from '@/app/auth/authOptions';
+
 export async function PATCH(
   request: NextRequest, 
   { params }: { params: { id: string }}
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    if (!session) return NextResponse.json({}, { status: 401 });
+
     const body = await request.json();
     const validation = issueSchema.safeParse(body);
     
@@ -26,8 +32,8 @@ export async function PATCH(
       where: { id: issue.id },
       data: {
         title: body.title,
-        description: body.description
-      }
+        description: body.description,
+      },
     });
   
     return NextResponse.json(updatedIssue, { status: 201 });
@@ -38,15 +44,19 @@ export async function PATCH(
 
 export async function DELETE(
   request: NextRequest, 
-  { params }: { params: { id: string }}
+  { params }: { params: { id: string } }
 ) {
   try {
+    const session = await getServerSession(authOptions);
+    
+    if (!session) return NextResponse.json({}, { status: 401 });
+
     const issue = await prisma.issue.findUnique({
       where: { id: parseInt(params.id) }
     });
   
     if (!issue)
-      return NextResponse.json({ error: 'Invalid issue'}, { status: 404 });
+      return NextResponse.json({ error: 'Invalid issue' }, { status: 404 });
   
     await prisma.issue.delete({
       where: { id: issue.id }
